Add javascript data.js file. Update index.html, desktop.css, mobile.css file
