+++ README.md
@@ -1,1 +1,39 @@
 # Tech-AI-Conference
+
+## Built With
+
+- Html
+- Css
+- Javascript
+
+## Author
+
+👤 **Emmanuel Effiong Essien**
+
+- GitHub: [@githubhandle](https://github.com/ehma90)
+- Twitter: [@twitterhandle](https://twitter.com/ehma_essien)
+
+
+## 🤝  Design Reference
+
+👤 **Cindy Shin (GUI & Graphic Designer)**
+
+- Behance: [@adagio07](https://www.behance.net/adagio07)
+
+## Github Page
+
+Here is a link to my github page https://ehma90.github.io/Tech-AI-Conference/
+
+## Show your support
+
+Give a ⭐ if you like this project!
+
+## Acknowledgments
+
+- Hat tip to anyone whose code was used
+- Inspiration
+- etc
+
+## 📝 License
+
+This project is [MIT](./MIT.md) licensed.
