+++ about.html
@@ -6,9 +6,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.css">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/css/bootstrap.min.css" integrity="sha384-xOolHFLEh07PJGoPkLv1IbcEPTNtaed2xpHsD9ESMhqIYd0nLMwNLD69Npy4HI+N" crossorigin="anonymous">
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="style1.css">
-    <script defer src="main.js"></script>
+    <link rel="stylesheet" href="mobile.css">
+    <link rel="stylesheet" href="desktop.css">
     <title>Tech-AI-Conference</title>
 </head>
 <body>
@@ -30,7 +29,7 @@
 
             <ul class="navlist">
                 <li><i class="fa fa-times-circle-o close" aria-hidden="true"></i></li>
-                <li><a href="about.html" class="nav-item">About</a></li>
+                <li><a href="index.html" class="nav-item">About</a></li>
                 <li><a href="#" class="nav-item">Program</a></li>
                 <li><a href="#" class="nav-item">Join</a></li>
                 <li><a href="#" class="nav-item">Sponsor</a></li>
@@ -55,15 +54,11 @@
     </section>
 
     <section id="about-logo">
-        <h2 class="about-header">Technological Advancement <br> in AI </h2>
+        <h2 class="about-header">Technological Advancement <br> Logo</h2>
         <div class="bar"></div>
         <p class="about-text">
-            Lorem ipsum, dolor sit amet consectetur adipisicing elit.
-             Illum corporis optio voluptatem ab autem tenetur unde pariatur 
-             quia in eius quod aliquid,Lorem ipsum dolor sit amet consectetur
-              adipisicing elit. Veniam dignissimos eum, nisi amet totam accusantium 
-              officiis. Iste perspiciatis, aperiam ipsam consequatur consequuntur dolorum 
-              accusamus temporibus, sunt porro veritatis quas at.
+            The logo of this company was decided through a logo competition from 3 April 2017. After online public voting the selection 
+            committee's pick a winner, and Emmanuel Essien won the prize. After much consultation with the board, we decide that this year's logo should be this:
         </p>
         <strong class="about-logo about-log">
             <i class="fa fa-snowflake-o" aria-hidden="true"></i><span>Ehma</span>Tech
@@ -109,6 +104,7 @@
         </p>
     </footer>
 
+    <script src="main.js" type="module"></script>
 
     <script src="https://cdn.jsdelivr.net/npm/jquery@3.5.1/dist/jquery.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-Fy6S3B9q64WdZWQUiU+q4/2Lc9npb8tCaSX9FK7E8HnRr0Jz8D6OP9dO5Vg3Q9ct" crossorigin="anonymous"></script>
