+++ data.js
@@ -1,0 +1,46 @@
+const data = [
+  {
+    id: 1,
+    image: 'imgs/person1.jpeg',
+    name: 'Alex Timber',
+    position: 'Lead AI engineer at AirBNB',
+    experience: 'He structures and coordinates the work of his/her collaborators who design and test sophisticated algorithms to carry out artificial intelligence projects.',
+  },
+  {
+    id: 2,
+    image: 'imgs/person2.jpeg',
+    name: 'Chou Zhao',
+    position: 'AI Engineer',
+    experience: 'Their role is to create machines capable of reasoning like the human brain.',
+  },
+  {
+    id: 3,
+    image: 'imgs/person3.jpeg',
+    name: 'John Lake',
+    position: 'Computer software',
+    experience: 'Write, test, debug, and maintain the detailed instructions, called computer programs, that computers must follow to perform their functions. Programmers also conceive, design, and test logical structures for solving problems by computer.',
+  },
+  {
+    id: 4,
+    image: 'imgs/person4.jpeg',
+    name: 'Sarah James',
+    position: 'Software  Engineer',
+    experience: 'Works on the technical team at an organization, typically under a senior management-level team member, to repair, develop, and maintain company software to ensure an efficient user experience.',
+  },
+  {
+    id: 5,
+    image: 'imgs/person5.jpeg',
+    name: 'Mark Benson',
+    position: 'Lead AI Engineer at Microsoft Inc',
+    experience: 'Use Cognitive Services, Machine Learning, and Knowledge Mining to architect and implement Microsoft AI solutions.',
+  },
+  {
+    id: 6,
+    image: 'imgs/person6.jpeg',
+    name: 'Jones Paul',
+    position: 'Senior AI developer',
+    experience: 'Build AI functionality into software applications. The role is generally focused on integrating and implementing AI algorithms and logic into the deliverables of an IT project. ',
+  },
+];
+
+export default data;
